Create RLE1.cs
Develop an algorithm which takes an input like 09a18b and outputs 9 as and 18 bs. The algorithm should use the first and second character of each data frequency pair  and parse into an integer so that it can be used in the loop as part of the condition.
There will only be inputs of 2 data frequency pairs each time
To store the 0 and the 9 into a separate string you will have to Convert.ToString(str[0] and [1] as it will not allow you to store the char directly.